speedy big decimal from decimal
